[chaturbate] add a description of the assumed logic behind cookie password saving
This was reverse engineered using multiple manual tests.
The cookie_password is only overwritten when a new valid password is sent. If the video_id_password is changed, the cookies will fail to log in until a valid password is sent again or the password goes back to the old one.